refactor: split api/auth/router.py into per-flow routers (CHAOS-1540)
Split the 1702-line monolithic router.py into per-flow sub-routers under
src/dev_health_ops/api/auth/routers/ following the admin routers pattern.

New structure:
- router.py        (52 lines)   thin aggregator, re-exports public symbols
- routers/__init__.py            exports all sub-routers + helpers
- routers/common.py (177 lines) shared models (UserInfo, LoginResponse,
                                 VerifyEmailResponse) + utility functions
                                 (_require_uuid, _parse_uuid, _expiry_to_utc,
                                 _resolve_login_audit_org_id,
                                 _issue_membership_tokens,
                                 _extract_unverified_org_and_subject, etc.)
- routers/dependencies.py (118) get_current_user, get_current_user_optional
- routers/register.py     (173) POST /register
- routers/verify.py       (100) GET /verify, POST /resend-verification
- routers/password_reset.py (95) POST /forgot-password, POST /reset-password
- routers/login.py        (320) POST /login
- routers/invites.py      (300) POST /accept-invite, POST /onboard
- routers/refresh.py      (217) POST /refresh
- routers/session.py      (146) GET /me, POST /validate, POST /logout
- routers/oauth.py        (207) POST /social-login

Preserved backward-compat public import paths:
- from dev_health_ops.api.auth.router import router (unchanged)
- from dev_health_ops.api.auth.router import get_current_user (re-exported)
- from dev_health_ops.api.auth.router import get_current_user_optional (re-exported)
- from dev_health_ops.api.auth.router import _extract_unverified_org_and_subject (re-exported)
- get_postgres_session, get_auth_service, create_refresh_token_record remain as
  module attributes on router.py to support existing test patching patterns
