+++ src/dev_health_ops/api/auth/routers/__init__.py
@@ -1,0 +1,48 @@
+from .common import (
+    LoginResponse,
+    UserInfo,
+    VerifyEmailResponse,
+    _coerce_uuid,
+    _expiry_to_utc,
+    _extract_unverified_org_and_subject,
+    _issue_membership_tokens,
+    _optional_uuid,
+    _parse_uuid,
+    _require_uuid,
+    _resolve_login_audit_org_id,
+    _slugify_org_name,
+)
+from .dependencies import get_current_user, get_current_user_optional
+from .invites import router as invites_router
+from .login import router as login_router
+from .oauth import router as oauth_router
+from .password_reset import router as password_reset_router
+from .refresh import router as refresh_router
+from .register import router as register_router
+from .session import router as session_router
+from .verify import router as verify_router
+
+__all__ = [
+    "LoginResponse",
+    "UserInfo",
+    "VerifyEmailResponse",
+    "_coerce_uuid",
+    "_expiry_to_utc",
+    "_extract_unverified_org_and_subject",
+    "_issue_membership_tokens",
+    "_optional_uuid",
+    "_parse_uuid",
+    "_require_uuid",
+    "_resolve_login_audit_org_id",
+    "_slugify_org_name",
+    "get_current_user",
+    "get_current_user_optional",
+    "invites_router",
+    "login_router",
+    "oauth_router",
+    "password_reset_router",
+    "refresh_router",
+    "register_router",
+    "session_router",
+    "verify_router",
+]
